refactor: divide fixture zome into integrity and coordinator
