+++ build-fixture.sh
@@ -4,9 +4,11 @@ set -e
 export RUSTFLAGS="-C opt-level=z --cfg getrandom_backend=\"custom\""
 
 # build fixture
-cd test/e2e/fixture/zomes/foo
+cd test/e2e/fixture/zomes/integrity/foo
 cargo build --release --target wasm32-unknown-unknown
-cd ../../
+cd ../../coordinator/foo
+cargo build --release --target wasm32-unknown-unknown
+cd ../../../
 hc dna pack . -o test.dna
 hc app pack . -o test.happ
 echo "Built fixture"
