+++ test/e2e/fixture/zomes/coordinator/foo/src/lib.rs
@@ -1,36 +1,5 @@
-use hdk::prelude::{holo_hash::DnaHash, *};
-
-#[hdk_entry_helper]
-pub struct TestString(pub String);
-
-#[derive(Debug, Serialize, Deserialize)]
-pub enum SerializationEnum {
-    Input,
-    Output(String),
-}
-
-impl From<String> for TestString {
-    fn from(s: String) -> Self {
-        Self(s)
-    }
-}
-
-impl From<&str> for TestString {
-    fn from(s: &str) -> Self {
-        Self(s.to_owned())
-    }
-}
-
-#[hdk_entry_types]
-#[unit_enum(UnitEntryTypes)]
-enum EntryTypes {
-    Test(TestString),
-}
-
-#[hdk_link_types]
-enum LinkTypes {
-    A,
-}
+use hdk::prelude::*;
+use test_wasm_foo_integrity::*;
 
 #[hdk_extern]
 fn init() -> ExternResult<InitCallbackResult> {
